Make lang.ast.Result a slimmer base class, extract code generation
